fix issue when user enters an invalid location when making a website. Before: the website would not get added to the swiftdata
Now:
when user enters an invalid locaiton, they are prompted with a popup.
The popup asks if wants to change the locaiton or continue.
when user presses continue the locaiton will be set to ""

Additionally renamed functions.
The code will now attemptSave, when figuring out if entered location is valid.
if location is valid or users presses continue, the save function is called.